+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - [UNRELEASED]
+## [0.4.0] - [2020-03-10]
 ### Added
+- EligibilityBenefit(s) Deductible information
+- EligibilityBenefit(s) Employee information
+- EligibilityData plan date information
+- EligibilityData eligibility, plan and service date helpers on date info
 
 ## [0.3.0] - [2020-03-09]
 ### Added
@@ -29,7 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Authentication
 - Configuration
 
-[0.4.0]: https://github.com/WeInfuse/change_health/compare/v0.3.0...HEAD
+[0.4.0]: https://github.com/WeInfuse/change_health/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/WeInfuse/change_health/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/WeInfuse/change_health/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/WeInfuse/change_health/compare/v0.1.0